refactor: move base to vectors
