+++ src/coordinax/_src/d2/base.py
@@ -5,13 +5,13 @@ __all__ = ["AbstractPos2D", "AbstractVel2D", "AbstractAcc2D"]
 
 from abc import abstractmethod
 
-from coordinax._src.base import (
+from coordinax._src.utils import classproperty
+from coordinax._src.vectors.base import (
     AbstractAcc,
     AbstractPos,
     AbstractVector,
     AbstractVel,
 )
-from coordinax._src.utils import classproperty
 
 
 class AbstractPos2D(AbstractPos):
