+++ src/coordinax/_src/d2/cartesian.py
@@ -22,9 +22,9 @@ from unxt.quantity import AbstractQuantity, Quantity
 
 import coordinax._src.typing as ct
 from .base import AbstractAcc2D, AbstractPos2D, AbstractVel2D
-from coordinax._src.base import AbstractPos
-from coordinax._src.base.mixins import AvalMixin
 from coordinax._src.utils import classproperty
+from coordinax._src.vectors.base import AbstractPos
+from coordinax._src.vectors.base.mixins import AvalMixin
 
 
 @final
