+++ src/coordinax/_src/d3/cartesian.py
@@ -25,9 +25,9 @@ from unxt.quantity import AbstractQuantity, Quantity
 import coordinax._src.typing as ct
 from .base import AbstractAcc3D, AbstractPos3D, AbstractVel3D
 from .generic import CartesianGeneric3D
-from coordinax._src.base import AbstractPos
-from coordinax._src.base.mixins import AvalMixin
 from coordinax._src.utils import classproperty
+from coordinax._src.vectors.base import AbstractPos
+from coordinax._src.vectors.base.mixins import AvalMixin
 
 #####################################################################
 # Position
