+++ src/coordinax/_src/dn/cartesian.py
@@ -19,9 +19,9 @@ from unxt import Quantity
 
 import coordinax._src.typing as ct
 from .base import AbstractAccND, AbstractPosND, AbstractVelND
-from coordinax._src.base import AbstractPos
-from coordinax._src.base.mixins import AvalMixin
 from coordinax._src.utils import classproperty
+from coordinax._src.vectors.base import AbstractPos
+from coordinax._src.vectors.base.mixins import AvalMixin
 
 ##############################################################################
 # Position
