+++ src/coordinax/_src/vectors/__init__.py
@@ -1,0 +1,3 @@
+"""Coordinax vectors."""
+
+__all__: list[str] = []
