+++ src/coordinax/_src/vectors/base/attribute.py
