+++ src/coordinax/_src/vectors/base/compat.py
