+++ src/coordinax/_src/vectors/base/flags.py
